backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
qdrant-client[async]==1.9.1
httpx==0.27.0
jinja2==3.1.4
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openpyxl==3.1.2
celery==5.4.0
anthropic==0.28.0
openai==1.30.1
numpy==1.26.4
pyyaml==6.0.1
